make use of heroku WEB_CONCURRENCY and do a cluster thing
